Align with existing API endpoints
