+++ functions/api/helpers.js
@@ -109,50 +109,47 @@ export async function refreshKPILatestValue(kpiRef) {
 }
 
 /**
- * Get list of KPIs including latest progress measurement.
+ * Build and return KPI response object.
+ *
+ * `kpiSnapshot` is the Firestore KPI document snapshot.
  */
-export async function getKPIs() {
-  const db = getFirestore();
+export async function buildKpiResponse(kpiSnapshot) {
+  const {
+    created,
+    createdBy,
+    currentValue,
+    edited,
+    editedBy,
+    name,
+    type,
+    updateFrequency,
+  } = kpiSnapshot.data();
 
-  const kpisSnapshot = await db
-    .collection('kpis')
-    .where('archived', '==', false)
-    .get();
-
-  const kpis = [];
-
-  for await (const kpiSnapshot of kpisSnapshot.docs) {
-    const { name, parent, valid, updateFrequency } = kpiSnapshot.data();
-
-    const parentName = await parent
-      .get()
-      .then((snapshot) => (snapshot.exists ? snapshot.data().name : null));
-
-    const latestMeasurement = await db
-      .collection(`kpis/${kpiSnapshot.id}/progress`)
-      .orderBy('timestamp', 'desc')
-      .limit(1)
-      .get()
-      .then((snapshot) => {
-        if (!snapshot.docs[0]) return null;
-        const { value, timestamp } = snapshot.docs[0].data();
-        return { value, timestamp: timestamp.toDate() };
-      });
-
-    const isStale = determineKPIStaleState(updateFrequency, latestMeasurement);
-
-    kpis.push({
-      id: kpiSnapshot.id,
-      name,
-      parentName,
-      valid,
-      latestMeasurement,
-      updateFrequency: updateFrequency || null,
-      isStale,
+  const latestMeasurement = await kpiSnapshot.ref
+    .collection('progress')
+    .orderBy('timestamp', 'desc')
+    .limit(1)
+    .get()
+    .then((snapshot) => {
+      if (!snapshot.docs[0]) return null;
+      const { value, timestamp } = snapshot.docs[0].data();
+      return { value, timestamp: timestamp.toDate() };
     });
-  }
 
-  return kpis;
+  const isStale = determineKPIStaleState(updateFrequency, latestMeasurement);
+
+  return {
+    currentValue,
+    name,
+    type,
+    lastUpdated: latestMeasurement || null,
+    updateFrequency: updateFrequency || null,
+    isStale,
+    created: created ? created.toDate() : null,
+    createdBy: createdBy ? await getUserDisplayName(createdBy) : null,
+    edited: edited ? edited.toDate() : null,
+    editedBy: editedBy ? await getUserDisplayName(editedBy) : null,
+  };
 }
 
 /**
